converted body to view
