+++ app/mobile/lib/modules/authentication/controllers/authentication_controller.dart
@@ -1,9 +1,11 @@
 import 'package:get/get.dart';
 import 'package:mobile/modules/authentication/views/auth_view.dart';
 
+
 class AuthenticationController extends GetxController {
   var isLogin = true.obs;
   var isForgotPassword = false.obs;
+
 
   void navigateToAuth({required bool toLogin}) {
     isLogin.value = toLogin;
