+++ app/mobile/lib/modules/authentication/controllers/forgot_password_controller.dart
@@ -1,5 +1,4 @@
 import 'package:get/get.dart';
-import 'package:get_storage/get_storage.dart';
 import 'package:mobile/data/helpers/validator.dart';
 
 import 'authentication_controller.dart';
@@ -10,29 +9,23 @@ class ForgotPasswordController extends GetxController {
   var verificationFailed = false.obs;
 
   final AuthenticationController authController =
-  Get.find<AuthenticationController>();
+      Get.find<AuthenticationController>();
 
   void onChangeUsername(String value) {
     loginEmail.value = value;
     loginEmailValid.value = Validator.isEmailValid(loginEmail.value);
-
-  }
-
-  void navigateToLogin() {
-    authController.toggleForgotPassword();
   }
 
   void onSubmit() async {
     // Forgot password submit logic here, then navigate to login
 
     // verify email, to see if it exists
-    if(true){
+    if (true) {
       verificationFailed = false.obs;
-      navigateToLogin();
-    }else{
+      Get.back();
+    } else {
       verificationFailed = true.obs;
     }
-
   }
 
   @override
