+++ app/mobile/lib/modules/authentication/views/auth_view.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mobile/data/constants/palette.dart';
 import 'package:mobile/modules/authentication/controllers/authentication_controller.dart';
-import 'package:mobile/modules/authentication/views/forgot_password_body.dart';
 import 'package:mobile/modules/authentication/views/login_body.dart';
 import 'package:mobile/modules/authentication/views/sign_up_body.dart';
 
@@ -18,11 +17,9 @@ class AuthView extends GetView<AuthenticationController> {
           if (controller.isLogin.value) {
             return const LoginBody();
           } else {
-            if(controller.isForgotPassword.value){
-              return const ForgotPasswordBody();
-            }else{
+            
               return const SignUpBody();
-            }
+            
           }
         }));
   }
